Music Genre Classification/pyproject.toml
[project]
name = "the-greatest-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "scikit-learn>=1.5.2",
    "flask>=3.1.0",
    "joblib>=1.4.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "spotipy>=2.24.0",
    "torch>=2.5.1",
    "xgboost>=2.1.2",
]
